backend/requirements.txt
Flask==2.0.1
flask-jwt-extended==4.3.1
flask-cors==3.0.10
python-dotenv==0.19.0
gunicorn==20.1.0
pymongo[srv]==4.6.0
Werkzeug==2.0.1
certifi==2023.7.22